+++ testing/src/test/scala/io/github/plume/oss/querying/MethodParameterTests.scala
@@ -20,14 +20,14 @@ class MethodParameterTests extends PlumeCodeToCpgSuite {
     val List(x) = cpg.parameter.name("param1").l
     x.code shouldBe "int param1"
     x.typeFullName shouldBe "int"
-    x.lineNumber shouldBe Some(2)
+    x.lineNumber shouldBe Some(3)
     // x.columnNumber shouldBe Some(11)
     x.order shouldBe 1
 
     val List(y) = cpg.parameter.name("param2").l
     y.code shouldBe "int param2"
     y.typeFullName shouldBe "int"
-    y.lineNumber shouldBe Some(2)
+    y.lineNumber shouldBe Some(3)
     // y.columnNumber shouldBe Some(21)
     y.order shouldBe 2
   }