+++ CHANGELOG.md
@@ -5,6 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.4] - 2021-03-15
+
+### Changed
+- Upped the default chunk size
+- `DeltaGraph::toOverflowDb` can now take in an optional `overflowdb.Graph` object to write to
+
+### Fixed
+- Memory leak where thread pools weren't getting shutdown
+
+## [0.3.3] - 2021-03-15
+
+### Added
+
+- `DeltaGraph` as a `NewNodeBuilder` variant of ShiftLeft's `DiffGraph`.
+- `BaseCpgPass` which is a combination of the `ASTPass`, `CFGPass`, and `PDGPass` and returns a `DeltaGraph` instead of
+  directly apply changes to the driver.
+- `methodBodies` was added to `GlobalCache` to save on database requests when moving to `SCPGPass` after `BaseCpgPass`
+- Chunk size can now be configured via `ExtractorOptions::methodChunkSize`
+
+### Changed
+
+- Replaced `ASTPass`, `CFGPass`, and `PDGPass` with `BaseCpgPass`.
+- Spawns a thread pool to run base CPG building in parallel and apply `DeltaGraph`s in serial.
+- SCPG flows are only run on new/updated method bodies since the analysis is independent of other methods.
+
 ## [0.3.2] - 2021-03-12
 
 ### Added
@@ -15,7 +40,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Moved the maps in `Extractor` to a dedicated `GlobalCache` object that uses `ConcurrentHashMap`s.
-- SCPG pass now concurrently pulls all methods and merges it into an input graph. This code has been moved to 
+- SCPG pass now concurrently pulls all methods and merges it into an input graph. This code has been moved to
   `passes.SCPGPass.kt`
 - External method stubs have call-to-returns generated i.e. (METHOD)-CFG->(RETURN)-CFG->(METHOD_RETURN)